task: make model modal its own component
